Move model into map folder
