+++ source/CMakeLists.txt
@@ -47,7 +47,7 @@ set(RADIX_SOURCES
   ${D}/map/XmlHelper.cpp
   ${D}/map/XmlMapLoader.cpp
   ${D}/map/material/MaterialLoader.cpp
-  ${D}/model/MeshLoader.cpp
+  ${D}/map/model/MeshLoader.cpp
   ${D}/physics/CollisionDispatcher.cpp
   ${D}/physics/KinematicCharacterController.cpp
   ${D}/physics/PhysicsDebugDraw.cpp
