+++ source/text/FontLoader.cpp
@@ -3,7 +3,7 @@
 #include <fstream>
 #include <sstream>
 
-#include <radix/model/MeshLoader.hpp>
+#include <radix/map/model/MeshLoader.hpp>
 #include <radix/texture/TextureLoader.hpp>
 #include <radix/env/Environment.hpp>
